+++ itests/diskscan/test3/test.sh
@@ -2,20 +2,22 @@
 
 set -eu
 
+BIN=../../../src/diskscan/ctguard-diskscan
+if ! [ -e ${BIN} ]; then
+    echo "Could not find binary at '${BIN}'!"
+    exit 1
+fi
+
 umask 0022
-
-SCRIPT=$(readlink -f "$0")
-SCRIPTPATH=$(dirname "$SCRIPT")
-cd "$SCRIPTPATH"
-
-echo "TEST #3 START"
 
 cleanup () {
   rm -Rf test_dir/
-  rm -f test.db test.output test5.txt test_diff.db
+  rm -f test.db test.output test.output.sorted test5.txt test_diff.db
 }
 
 cleanup
+
+chmod 640 test.conf
 
 mkdir test_dir/
 cat > test_dir/txt1.txt <<EOF
@@ -77,11 +79,7 @@ ln -s dead test_dir/lnk2
 rm test_dir/dead
 touch test_dir/test_rm
 
-if [ ! -x ../../../ctguard-diskscan ]; then
-	echo "TEST #3 Can not find executable!"
-	exit 1
-fi
-../../../ctguard-diskscan -c test.conf -x -f -S
+${BIN} -c test.conf -x -f -S
 
 cat > test_dir/txt1.txt <<EOF
 test file
@@ -167,11 +165,13 @@ ln -s test5.txt test_dir/lnk3
 rm test_dir/lnk1
 rm test_dir/test_rm
 
-../../../ctguard-diskscan -c test.conf -x -f -S
+${BIN} -c test.conf -x -f -S
 
-echo "TEST #3 Comapring expected vs actual output"
-diff -u test.expected test.output
+echo "Comapring expected vs actual output:"
+sed "s/ITEST_USER_REPLACEME/${USER}/g" test.expected.sample > test.expected
+sort test.output -o test.output.sorted
+diff -u test.expected test.output.sorted
 
 cleanup
 
-echo "TEST #3 success"
+echo "SUCCESS!"
